+++ libs/platform/core/src/services/query/core/build-event.ts
