+++ libs/platform/core/src/services/query/core/core-command.ts
@@ -16,7 +16,7 @@ import {
   CommandStrategy,
   QueryEventHandler,
 } from '../models';
-import { buildEvent } from './dispatch-event';
+import { buildEvent } from './build-event';
 
 export class CoreCommand<
   ResultType,
